+++ docs/app starting.txt
@@ -1,0 +1,53 @@
+
+
+~/ollama/bin/ollama serve
+
+uvicorn app.main:app --host 0.0.0.0 --port 6006 --reload
+
+curl.exe http://im-redstone02.hs-regensburg.de:32311/number
+
+curl.exe -X POST \
+  "http://im-redstone02.hs-regensburg.de:32311/process" \
+  -H "Content-Type: application/json" \
+  -d "{\"url\":\"https://www.instagram.com/reels/DIRM85ZifdM/\",\"mock\":\"true\"}"
+
+WINDOWS FUNKTIONIERT:
+curl "http://im-redstone02.hs-regensburg.de:32311/process" `
+     -Method Post `
+     -Body ( @{ url = "https://www.instagram.com/reels/DIRM85ZifdM/"; mock = "true" } | ConvertTo-Json ) `
+     -ContentType "application/json"
+
+
+
+curl -X POST http://localhost:6001/process \
+  -H "Content-Type: application/json" \
+  -d '{"url":"https://www.instagram.com/reels/DIRM85ZifdM/","mock":"true"}'
+
+
+https://www.instagram.com/reels/DIRM85ZifdM/  Butter
+
+https://www.instagram.com/reels/DI34PNII2iZ/  Drinking Prep
+https://www.instagram.com/reels/DIL8qejowvX/  OPC
+https://www.instagram.com/reels/DIlxIbpAjtK/  Beathing
+https://www.instagram.com/reels/DJDzRAgNHyv/  Entgiftung
+
+
+https://www.instagram.com/reels/DFycUIjy1C4/ 
+
+
+
+Modelle von Huggingface ziehen: (Public)
+ollama run hf.co/QuantFactory/Meditron3-8B-GGUF
+ollama run huggingface.co/mradermacher/Meditron3-70B-GGUF
+
+Umbenennen: (Kopieren)
+ollama cp huggingface.co/mradermacher/Meditron3-70B-GGUF meditron3-70b
+
+Löschen:
+ollama rm huggingface.co/mradermacher/Meditron3-70B-GGUF
+
+Modelle auflisten:
+ollama ls
+
+python -m pip install -r requirements.txt
+conda env create -f environment.yml
