+++ evaluation/evaluation_BERT.py
@@ -21,7 +21,7 @@ import inspect
 import json
 import os
 import random
-from typing import Any, Dict, List, Optional, Tuple
+from typing import Any, Dict, List, Optional
 
 import numpy as np
 import torch
