+++ evaluation/evaluation_raw.py
@@ -1,18 +1,17 @@
 #!/usr/bin/env python3
-import json
-import time
 import copy
+import json
 import os
-import re
 import random
-from datetime import datetime, timezone
+import re
+import time
 from collections import Counter, defaultdict
+from datetime import datetime, timezone
 from typing import Optional, List, Dict, Any
 
-from app2.core.models import PipelineState
-from app2.core.orchestrator import PipelineOrchestrator
-import app2.test_configs.raw_eval_config as config_module
-
+import pipeline.test_configs.raw_eval_config as config_module
+from pipeline.core.models import PipelineState
+from pipeline.core.orchestrator import PipelineOrchestrator
 
 # Mapping: pipeline verdict -> dataset label
 # Order requested: true -> 0, uncertain -> 1, false -> 2
